pyproject.toml
[tool.poetry]
name = "business-card-ocr"
version = "0.1.0"
description = ""
authors = ["Edmund Situmorang <edmund@situmorang.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.12"
pytesseract = "^0.3.13"
opencv-python = "^4.10.0.84"
pillow = "^11.0.0"
pyheif = "^0.8.0"
spacy = "^3.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
